+++ TagsCloudVisualization/TagsCloudVisualization/TagsCloudVisualization/CloudParameters.cs
@@ -5,50 +5,31 @@ namespace TagsCloudVisualization
 {
 	public class CloudParameters
 	{
-		public static CloudParametersDTO Parse(string[] input)
-		{
-			var parameters = new CloudParametersDTO();
-			var count = 0;
-			var maxLengthRect = 0;
+		public int MaxLengthRect { get; set; }
+		public int Count { get; set; }
+		public ICurve Curve { get; set; }
 
-			for (var i = 0; i < input.Length; i++)
+		public bool IsCorrect()
+		{
+			if (Curve == null)
 			{
-				switch (input[i])
-				{
-					case "-count":
-						int.TryParse(input[i + 1], out count);
-						break;
-					case "-maxLength":
-						int.TryParse(input[i + 1], out maxLengthRect);
-						break;
-					case "-curve":
-						parameters.Curve = GetCurve(input, i);
-						break;
-				}
+				Console.WriteLine("Error in the name of the curve. You need to choose one of them: spiral | heart | astroid");
+				return false;
 			}
 
-			parameters.Count = count;
-			parameters.MaxLengthRect = maxLengthRect;
-			return parameters;
-		}
-
-		private static ICurve GetCurve(string[] args, int position)
-		{
-			ICurve curve = null;
-			switch (args[position + 1])
+			if (Count <= 0)
 			{
-				case "spiral":
-					curve = new Spiral(0.2, Math.PI / 36);
-					break;
-				case "heart":
-					curve = new Heart(0.2, Math.PI / 36);
-					break;
-				case "astroid":
-					curve = new Astroid(0.2, Math.PI / 36);
-					break;
+				Console.WriteLine("Count must be positive");
+				return false;
 			}
 
-			return curve;
+			if (MaxLengthRect <= 0)
+			{
+				Console.WriteLine("MaxLength must be positive");
+				return false;
+			}
+
+			return true;
 		}
 	}
 }
