+++ TagsCloudVisualization/TagsCloudVisualization/TagsCloudVisualization/TagsCloudVisualization.csproj
@@ -55,12 +55,12 @@
   <ItemGroup>
     <Compile Include="CircularCloudLayouter.cs" />
     <Compile Include="CircularCloudLayouterTests.cs" />
-    <Compile Include="CloudParametersDTO.cs" />
+    <Compile Include="CloudParameters.cs" />
     <Compile Include="Curves\Astroid.cs" />
     <Compile Include="Curves\Heart.cs" />
     <Compile Include="Curves\ICurve.cs" />
     <Compile Include="Curves\Spiral.cs" />
-    <Compile Include="CloudParameters.cs" />
+    <Compile Include="CloudParametersParser.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RectangleExtesions.cs" />
